+++ tauon-app/src/main/java/tauon/app/settings/Settings.java
@@ -51,7 +51,6 @@ public class Settings {
     private boolean terminalBell = false;
     private String terminalFontName = FontUtils.TERMINAL_FONTS.keySet().iterator().next();
     private int terminalFontSize = 14;
-    private Language language = Language.ENGLISH;
     private String terminalTheme = "Dark";
     private String terminalPalette = "xterm";
     private int[] palleteColors = {0x000000, 0xcd0000, 0x00cd00, 0xcdcd00, 0x1e90ff, 0xcd00cd, 0x00cdcd, 0xe5e5e5,
@@ -443,20 +442,6 @@ public class Settings {
      */
     public void setTerminalTheme(String terminalTheme) {
         this.terminalTheme = terminalTheme;
-    }
-
-    /**
-     * @return Language
-     */
-    public Language getLanguage() {
-        return language;
-    }
-
-    /**
-     * @param language
-     */
-    public void setLanguage(Language language) {
-        this.language = language;
     }
 
     /**